minimax changes

+++ src/shared/ai/Node.cpp
@@ -10,8 +10,8 @@ using namespace ai;
 
 Node::Node(state::State& currState) : char_p1(currState.getListCharacters(0)), char_p2(currState.getListCharacters(1))
 {
-    for (int i = 0; i < char_p1.size(); i++) value += char_p1[i]->getHealth();
-    for (int i = 0; i < char_p2.size(); i++) value -= char_p2[i]->getHealth();
+    for (size_t i = 0; i < char_p1.size(); i++) value += char_p1[i]->getHealth();
+    for (size_t i = 0; i < char_p2.size(); i++) value -= char_p2[i]->getHealth();
 }
